testnet support. inventory and asset history. trimmed unneeded route paths.
