+++ app.js
@@ -5,27 +5,37 @@ const path = require('path')
 const cookieParser = require('cookie-parser')
 const logger = require('morgan')
 
-//otp---
-
 //dkg
-const getRouter_otp = require('./routes/otp/dkg/get')
-const getLatestStateIssuerRouter_otp = require('./routes/otp/dkg/getLatestStateIssuer')
-const getOwnerRouter_otp = require('./routes/otp/dkg/getOwner')
-const getStateIssuerRouter_otp = require('./routes/otp/dkg/getStateIssuer')
-const getStatesRouter_otp = require('./routes/otp/dkg/getStates')
-const createRouter_otp = require('./routes/otp/dkg/create')
-const queryRouter_otp = require('./routes/otp/dkg/query')
-const transferRouter_otp = require('./routes/otp/dkg/transfer')
-const updateRouter_otp = require('./routes/otp/dkg/update')
+const getRouter_otp = require('./routes/dkg/get')
+const getLatestStateIssuerRouter_otp = require('./routes/dkg/getLatestStateIssuer')
+const getOwnerRouter_otp = require('./routes/dkg/getOwner')
+const getStateIssuerRouter_otp = require('./routes/dkg/getStateIssuer')
+const getStatesRouter_otp = require('./routes/dkg/getStates')
+const createRouter_otp = require('./routes/dkg/create')
+const queryRouter_otp = require('./routes/dkg/query')
+const transferRouter_otp = require('./routes/dkg/transfer')
+const create_n_transferRouter_otp = require('./routes/dkg/create_n_transfer')
+const updateRouter_otp = require('./routes/dkg/update')
 
-//views
-const v_nodes_statsRouter_otp = require('./routes/otp/views/v_nodes_stats')
-const v_nodes_stats_lastRouter_otp = require('./routes/otp/views/v_nodes_stats_last')
-const v_nodes_rank_lastRouter_otp = require('./routes/otp/views/v_nodes_rank_last')
-const v_nodesRouter_otp = require('./routes/otp/views/v_nodes')
+//otp
+const v_nodes_statsRouter_otp = require('./routes/otp/v_nodes_stats')
+const v_nodes_stats_lastRouter_otp = require('./routes/otp/v_nodes_stats_last')
+const v_nodes_rank_lastRouter_otp = require('./routes/otp/v_nodes_rank_last')
+const v_nodesRouter_otp = require('./routes/otp/v_nodes')
+const v_pubs_statsRouter_otp = require('./routes/otp/v_pubs_stats')
+const v_pubs_stats_lastRouter_otp = require('./routes/otp/v_pubs_stats_last')
+const assetHistory_otp = require('./routes/otp/assetHistory')
+const assetInventory_otp = require('./routes/otp/assetInventory')
 
-const v_pubs_statsRouter_otp = require('./routes/otp/views/v_pubs_stats')
-const v_pubs_stats_lastRouter_otp = require('./routes/otp/views/v_pubs_stats_last')
+//otp_testnet
+const v_nodes_statsRouter_otp_testnet = require('./routes/otp_testnet/v_nodes_stats')
+const v_nodes_stats_lastRouter_otp_testnet = require('./routes/otp_testnet/v_nodes_stats_last')
+const v_nodes_rank_lastRouter_otp_testnet = require('./routes/otp_testnet/v_nodes_rank_last')
+const v_nodesRouter_otp_testnet = require('./routes/otp_testnet/v_nodes')
+const v_pubs_statsRouter_otp_testnet = require('./routes/otp_testnet/v_pubs_stats')
+const v_pubs_stats_lastRouter_otp_testnet = require('./routes/otp_testnet/v_pubs_stats_last')
+const assetHistory_otp_testnet = require('./routes/otp_testnet/assetHistory')
+const assetInventory_otp_testnet = require('./routes/otp_testnet/assetInventory')
 
 const app = express()
 
@@ -38,25 +48,37 @@ app.use(express.static(path.join(__dirname, 'node_modules')))
 app.use('/dkg', express.static(__dirname + 'node_modules/dkg.js'))
 app.use('/util', express.static(__dirname + 'public/util'))
 
-//api
-app.use('/otp/views/v_nodes_stats', v_nodes_statsRouter_otp)
-app.use('/otp/views/v_nodes_stats_last', v_nodes_stats_lastRouter_otp)
-app.use('/otp/views/v_nodes_rank_last', v_nodes_rank_lastRouter_otp)
-app.use('/otp/views/v_nodes', v_nodesRouter_otp)
-
-app.use('/otp/views/v_pubs_stats', v_pubs_statsRouter_otp)
-app.use('/otp/views/v_pubs_stats_last', v_pubs_stats_lastRouter_otp)
-
 //dkg
-app.use('/otp/dkg/get', getRouter_otp)
-app.use('/otp/dkg/getLatestStateIssuer', getLatestStateIssuerRouter_otp)
-app.use('/otp/dkg/getOwner', getOwnerRouter_otp)
-app.use('/otp/dkg/getStateIssuer', getStateIssuerRouter_otp)
-app.use('/otp/dkg/getStates', getStatesRouter_otp)
-app.use('/otp/dkg/create', createRouter_otp)
-app.use('/otp/dkg/update', updateRouter_otp)
-app.use('/otp/dkg/query', queryRouter_otp)
-app.use('/otp/dkg/transfer', transferRouter_otp)
+app.use('/dkg/get', getRouter_otp)
+app.use('/dkg/getLatestStateIssuer', getLatestStateIssuerRouter_otp)
+app.use('/dkg/getOwner', getOwnerRouter_otp)
+app.use('/dkg/getStateIssuer', getStateIssuerRouter_otp)
+app.use('/dkg/getStates', getStatesRouter_otp)
+app.use('/dkg/create', createRouter_otp)
+app.use('/dkg/update', updateRouter_otp)
+app.use('/dkg/query', queryRouter_otp)
+app.use('/dkg/transfer', transferRouter_otp)
+app.use('/dkg/create_n_transfer', create_n_transferRouter_otp)
+
+//otp
+app.use('/otp/v_nodes_stats', v_nodes_statsRouter_otp)
+app.use('/otp/v_nodes_stats_last', v_nodes_stats_lastRouter_otp)
+app.use('/otp/v_nodes_rank_last', v_nodes_rank_lastRouter_otp)
+app.use('/otp/v_nodes', v_nodesRouter_otp)
+app.use('/otp/v_pubs_stats', v_pubs_statsRouter_otp)
+app.use('/otp/v_pubs_stats_last', v_pubs_stats_lastRouter_otp)
+app.use('/otp/assetHistory', assetHistory_otp)
+app.use('/otp/assetInventory', assetInventory_otp)
+
+//otp_testnet
+app.use('/otp_testnet/v_nodes_stats', v_nodes_statsRouter_otp_testnet)
+app.use('/otp_testnet/v_nodes_stats_last', v_nodes_stats_lastRouter_otp_testnet)
+app.use('/otp_testnet/v_nodes_rank_last', v_nodes_rank_lastRouter_otp_testnet)
+app.use('/otp_testnet/v_nodes', v_nodesRouter_otp_testnet)
+app.use('/otp_testnet/v_pubs_stats', v_pubs_statsRouter_otp_testnet)
+app.use('/otp_testnet/v_pubs_stats_last', v_pubs_stats_lastRouter_otp_testnet)
+app.use('/otp_testnet/assetHistory', assetHistory_otp_testnet)
+app.use('/otp_testnet/assetInventory', assetInventory_otp_testnet)
 
 
 // catch 404 and forward to error handler
